fix(onnx): harden Gather and LayerNorm semantics

+++ include/kxc/relay/op.h
@@ -264,12 +264,12 @@ public:
     static SoftmaxAttrs Create(int axis);
 };
 
-/*! \brief nn.layer_norm axis, epsilon, and required accumulation dtype. */
+/*! \brief nn.layer_norm axis, float32 epsilon, and float64 internal accumulation. */
 class LayerNormAttrsNode : public BaseAttrsNode {
 public:
     int axis = -1;
     float epsilon = 1e-5f;
-    std::string accumulation_dtype = "float32";
+    std::string accumulation_dtype = "float64";
     void SerializeCanonical(CanonicalAttrWriter& writer) const override;
     KXC_DECLARE_ATTRS_NODE
 };
@@ -278,7 +278,7 @@ class LayerNormAttrs : public Attrs {
 
 public:
     static LayerNormAttrs Create(int axis = -1, float epsilon = 1e-5f,
-                                 std::string accumulation_dtype = "float32");
+                                 std::string accumulation_dtype = "float64");
 };
 
 KXC_DEFINE_SIMPLE_ATTRS(AddAttrs)
